PrIfCompiler + PrNotCompiler
