+++ crates/core/src/pattern/not.rs
@@ -1,27 +1,20 @@
 use super::{
     functions::{Evaluator, FuncEvaluation},
-    iter_pattern::PatternOrPredicate,
     patterns::{Matcher, Name, Pattern},
     predicates::Predicate,
     resolved_pattern::ResolvedPattern,
-    variable::VariableSourceLocations,
     State,
 };
-use crate::{
-    context::Context,
-    pattern_compiler::{predicate_compiler::PredicateCompiler, CompilationContext, NodeCompiler},
-};
-use anyhow::{anyhow, bail, Ok, Result};
+use crate::context::Context;
+use anyhow::{bail, Ok, Result};
 use core::fmt::Debug;
-use marzano_util::analysis_logs::{AnalysisLogBuilder, AnalysisLogs};
-use marzano_util::position::Range;
-use std::collections::BTreeMap;
-use tree_sitter::Node;
+use marzano_util::analysis_logs::AnalysisLogs;
 
 #[derive(Debug, Clone)]
 pub struct Not {
     pub pattern: Pattern,
 }
+
 impl Not {
     pub fn new(pattern: Pattern) -> Self {
         Self { pattern }
@@ -56,47 +49,6 @@ pub struct PrNot {
 impl PrNot {
     pub fn new(predicate: Predicate) -> Self {
         Self { predicate }
-    }
-    pub(crate) fn from_node(
-        node: &Node,
-        context: &CompilationContext,
-        vars: &mut BTreeMap<String, usize>,
-        vars_array: &mut Vec<Vec<VariableSourceLocations>>,
-        scope_index: usize,
-        global_vars: &mut BTreeMap<String, usize>,
-        logs: &mut AnalysisLogs,
-    ) -> Result<Self> {
-        let not = node
-            .child_by_field_name("predicate")
-            .ok_or_else(|| anyhow!("predicateNot missing predicate"))?;
-        let range: Range = not.range().into();
-        let not = PredicateCompiler::from_node(
-            &not,
-            context,
-            vars,
-            vars_array,
-            scope_index,
-            global_vars,
-            logs,
-        )?;
-        if not.iter().any(|p| {
-            matches!(
-                p,
-                PatternOrPredicate::Pattern(Pattern::Rewrite(_))
-                    | PatternOrPredicate::Predicate(Predicate::Rewrite(_))
-            )
-        }) {
-            let log = AnalysisLogBuilder::default()
-                .level(441_u16)
-                .file(context.file)
-                .source(context.src)
-                .position(range.start)
-                .range(range)
-                .message("Warning: rewrites inside of a not will never be applied")
-                .build()?;
-            logs.push(log);
-        }
-        Ok(PrNot::new(not))
     }
 }
