+++ crates/core/src/pattern_compiler/if_compiler.rs
@@ -2,7 +2,11 @@ use super::{
     compiler::CompilationContext, node_compiler::NodeCompiler,
     predicate_compiler::PredicateCompiler,
 };
-use crate::pattern::{patterns::Pattern, r#if::If, variable::VariableSourceLocations};
+use crate::pattern::{
+    patterns::Pattern,
+    r#if::{If, PrIf},
+    variable::VariableSourceLocations,
+};
 use anyhow::{anyhow, Result};
 use marzano_util::analysis_logs::AnalysisLogs;
 use std::collections::BTreeMap;
@@ -63,5 +67,61 @@ impl NodeCompiler for IfCompiler {
             })
             .map_or(Ok(None), |v| v.map(Some))?;
         Ok(If::new(if_, then, else_))
+    }
+}
+
+pub(crate) struct PrIfCompiler;
+
+impl NodeCompiler for PrIfCompiler {
+    type TargetPattern = PrIf;
+
+    fn from_node(
+        node: &Node,
+        context: &CompilationContext,
+        vars: &mut BTreeMap<String, usize>,
+        vars_array: &mut Vec<Vec<VariableSourceLocations>>,
+        scope_index: usize,
+        global_vars: &mut BTreeMap<String, usize>,
+        logs: &mut AnalysisLogs,
+    ) -> Result<Self::TargetPattern> {
+        let if_ = node
+            .child_by_field_name("if")
+            .ok_or_else(|| anyhow!("missing condition of if"))?;
+        let if_ = PredicateCompiler::from_node(
+            &if_,
+            context,
+            vars,
+            vars_array,
+            scope_index,
+            global_vars,
+            logs,
+        )?;
+        let then = node
+            .child_by_field_name("then")
+            .ok_or_else(|| anyhow!("missing consequence of if"))?;
+        let then = PredicateCompiler::from_node(
+            &then,
+            context,
+            vars,
+            vars_array,
+            scope_index,
+            global_vars,
+            logs,
+        )?;
+        let else_ = node
+            .child_by_field_name("else")
+            .map(|e| {
+                PredicateCompiler::from_node(
+                    &e,
+                    context,
+                    vars,
+                    vars_array,
+                    scope_index,
+                    global_vars,
+                    logs,
+                )
+            })
+            .map_or(Ok(None), |v| v.map(Some))?;
+        Ok(PrIf::new(if_, then, else_))
     }
 }
