+++ crates/core/src/pattern_compiler/not_compiler.rs
@@ -1,6 +1,12 @@
-use super::{compiler::CompilationContext, node_compiler::NodeCompiler};
+use super::{
+    compiler::CompilationContext, node_compiler::NodeCompiler,
+    predicate_compiler::PredicateCompiler,
+};
 use crate::pattern::{
-    iter_pattern::PatternOrPredicate, not::Not, patterns::Pattern, predicates::Predicate,
+    iter_pattern::PatternOrPredicate,
+    not::{Not, PrNot},
+    patterns::Pattern,
+    predicates::Predicate,
     variable::VariableSourceLocations,
 };
 use anyhow::{anyhow, Result};
@@ -57,5 +63,53 @@ impl NodeCompiler for NotCompiler {
             logs.push(log);
         }
         Ok(Not::new(pattern))
+    }
+}
+
+pub(crate) struct PrNotCompiler;
+
+impl NodeCompiler for PrNotCompiler {
+    type TargetPattern = PrNot;
+
+    fn from_node(
+        node: &Node,
+        context: &CompilationContext,
+        vars: &mut BTreeMap<String, usize>,
+        vars_array: &mut Vec<Vec<VariableSourceLocations>>,
+        scope_index: usize,
+        global_vars: &mut BTreeMap<String, usize>,
+        logs: &mut AnalysisLogs,
+    ) -> Result<Self::TargetPattern> {
+        let not = node
+            .child_by_field_name("predicate")
+            .ok_or_else(|| anyhow!("predicateNot missing predicate"))?;
+        let range: Range = not.range().into();
+        let not = PredicateCompiler::from_node(
+            &not,
+            context,
+            vars,
+            vars_array,
+            scope_index,
+            global_vars,
+            logs,
+        )?;
+        if not.iter().any(|p| {
+            matches!(
+                p,
+                PatternOrPredicate::Pattern(Pattern::Rewrite(_))
+                    | PatternOrPredicate::Predicate(Predicate::Rewrite(_))
+            )
+        }) {
+            let log = AnalysisLogBuilder::default()
+                .level(441_u16)
+                .file(context.file)
+                .source(context.src)
+                .position(range.start)
+                .range(range)
+                .message("Warning: rewrites inside of a not will never be applied")
+                .build()?;
+            logs.push(log);
+        }
+        Ok(PrNot::new(not))
     }
 }
